Add match only to session or system buses
DBus connections that are not session nor system buses do not implement
`AddMatch`. So, trying to add a match in `ldbus.api.watch` results in
the following error being returned by DBus:

`Method "AddMatch" with signature "s" on interface
"org.freedesktop.DBus" doesn't exist`

To fix the issue, simply skip the step to add a match.

Manually verified to work with PulseAudio. Adding tests for this seems
too painful compared with the benefit received.

References:

* https://github.com/mvidner/ruby-dbus/issues/44